requirements.txt
Pillow
aiohttp
discord.py
requests
flask
flask-cors